Safari-611.1.10/LayoutTests/fast/misc/test-observegc-expected.txt

Ensures that window.internals.observegc works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS anyCollected is true
PASS successfullyParsed is true

TEST COMPLETE
